* Try using fmodf() instead
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2544 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/UI/GuiHandler.cpp
@@ -3564,7 +3564,7 @@ void CGuiHandler::DrawMiniMapCursor()
 	const float groundLevel = ground->GetHeight(camera->pos.x, camera->pos.z);
 
 	static float spinTime = 0.0f;
-	spinTime = (float)fmod(spinTime + gu->lastFrameTime, 60.0);
+	spinTime = fmodf(spinTime + gu->lastFrameTime, 60.0f);
 
 	glPushMatrix();
 	glTranslatef(camera->pos.x, groundLevel, camera->pos.z);
